backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
pymongo==4.10.1
python-dotenv==1.0.0
Werkzeug==3.0.1
requests==2.31.0
google-generativeai==0.8.3
gunicorn==21.2.0
